dockerProject/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.30
Werkzeug==2.2.2
Jinja2==3.1.2